requirements.txt
streamlit
streamlit-extras
streamlit-option-menu
pytest
openai
langchain
trufflehog
PyGithub
boto3
slack_sdk